#12 fixing the hourly updates and adding comments to my code.
In this commit I fixed the link that fetches information from open meteo to also include the hourly information (fixing a bug I was having previously which made our app not start the hourly updates from the current time - it went from 00:00 to 11:00 before). I also added extensive comments explainning what is being done in each section).

Next section is going to be removing some UI elements that are not being functionally used.