#14 Typescript fix
We kept getting an error in our code saying that our parameters did not have a defined type, I researched and found out that though our code was a valid java script code, and that is why it was working, it lacked a few details necessary in TypeScript. I adapted the code to define types to our parameters, fixing that error.